Moved most of the remaining classes.
Still have some more to migrate, but is enough to start progress on
other steps.
