+++ amazon-kinesis-client/src/main/java/software/amazon/kinesis/coordinator/KinesisClientLibConfiguration.java
@@ -12,12 +12,14 @@
  *  express or implied. See the License for the specific language governing
  *  permissions and limitations under the License. 
  */
-package com.amazonaws.services.kinesis.clientlibrary.lib.worker;
+package software.amazon.kinesis.coordinator;
 
 import java.util.Date;
 import java.util.Optional;
 import java.util.Set;
 
+import com.amazonaws.services.kinesis.clientlibrary.lib.worker.InitialPositionInStream;
+import com.amazonaws.services.kinesis.clientlibrary.lib.worker.InitialPositionInStreamExtended;
 import org.apache.commons.lang.Validate;
 
 import com.amazonaws.ClientConfiguration;
