+++ amazon-kinesis-client/src/main/java/software/amazon/kinesis/coordinator/WorkerStateChangeListener.java
@@ -1,0 +1,30 @@
+/*
+ *  Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ *  Licensed under the Amazon Software License (the "License").
+ *  You may not use this file except in compliance with the License.
+ *  A copy of the License is located at
+ *
+ *  http://aws.amazon.com/asl/
+ *
+ *  or in the "license" file accompanying this file. This file is distributed
+ *  on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ *  express or implied. See the License for the specific language governing
+ *  permissions and limitations under the License.
+ */
+package software.amazon.kinesis.coordinator;
+
+/**
+ * A listener for callbacks on changes worker state
+ */
+@FunctionalInterface
+public interface WorkerStateChangeListener {
+	enum WorkerState {
+		CREATED,
+		INITIALIZING,
+		STARTED,
+		SHUT_DOWN
+	}
+
+	void onWorkerStateChange(WorkerState newState);
+}
