+++ amazon-kinesis-client/src/main/java/software/amazon/kinesis/lifecycle/ShutdownNotification.java
@@ -12,7 +12,7 @@
  *  express or implied. See the License for the specific language governing
  *  permissions and limitations under the License.
  */
-package com.amazonaws.services.kinesis.clientlibrary.lib.worker;
+package software.amazon.kinesis.lifecycle;
 
 import software.amazon.kinesis.lifecycle.ShutdownInput;
 import software.amazon.kinesis.processor.v2.IRecordProcessor;
