+++ amazon-kinesis-client/src/main/java/software/amazon/kinesis/retrieval/KinesisProxyFactory.java
@@ -18,7 +18,7 @@ import com.amazonaws.ClientConfiguration;
 import com.amazonaws.auth.AWSCredentialsProvider;
 import com.amazonaws.services.kinesis.AmazonKinesis;
 import com.amazonaws.services.kinesis.AmazonKinesisClient;
-import com.amazonaws.services.kinesis.clientlibrary.lib.worker.KinesisClientLibConfiguration;
+import software.amazon.kinesis.coordinator.KinesisClientLibConfiguration;
 
 /**
  * Factory used for instantiating KinesisProxy objects (to fetch data from Kinesis).
