+++ amazon-kinesis-client/src/test/java/software/amazon/kinesis/utils/TestStreamletFactory.java
@@ -1,18 +1,18 @@
 /*
- * Copyright 2012-2016 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *  Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
- * Licensed under the Amazon Software License (the "License").
- * You may not use this file except in compliance with the License.
- * A copy of the License is located at
+ *  Licensed under the Amazon Software License (the "License").
+ *  You may not use this file except in compliance with the License.
+ *  A copy of the License is located at
  *
- * http://aws.amazon.com/asl/
+ *  http://aws.amazon.com/asl/
  *
- * or in the "license" file accompanying this file. This file is distributed
- * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing
- * permissions and limitations under the License.
+ *  or in the "license" file accompanying this file. This file is distributed
+ *  on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ *  express or implied. See the License for the specific language governing
+ *  permissions and limitations under the License.
  */
-package com.amazonaws.services.kinesis.clientlibrary.lib.worker;
+package software.amazon.kinesis.utils;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -25,7 +25,7 @@ import software.amazon.kinesis.processor.v2.IRecordProcessorFactory;
 /**
  * Factory for TestStreamlet record processors.
  */
-class TestStreamletFactory implements IRecordProcessorFactory {
+public class TestStreamletFactory implements IRecordProcessorFactory {
 
     // Will be passed to the TestStreamlet. Can be used to check if all records have been processed.
     private Semaphore semaphore;
@@ -35,7 +35,7 @@ class TestStreamletFactory implements IRecordProcessorFactory {
     /**
      *  Constructor.
      */
-    TestStreamletFactory(Semaphore semaphore, ShardSequenceVerifier shardSequenceVerifier) {
+    public TestStreamletFactory(Semaphore semaphore, ShardSequenceVerifier shardSequenceVerifier) {
         this.semaphore = semaphore;
         this.shardSequenceVerifier = shardSequenceVerifier;
     }
@@ -51,14 +51,14 @@ class TestStreamletFactory implements IRecordProcessorFactory {
         return semaphore;
     }
 
-    ShardSequenceVerifier getShardSequenceVerifier() {
+    public ShardSequenceVerifier getShardSequenceVerifier() {
         return shardSequenceVerifier;
     }
 
     /**
      * @return the testStreamlets
      */
-    List<TestStreamlet> getTestStreamlets() {
+    public List<TestStreamlet> getTestStreamlets() {
         return testStreamlets;
     }
